web/requirements_ui.txt
flask==2.3.3
flask-socketio==5.3.6
python-dotenv==1.0.0
werkzeug==2.3.7
PyYAML==6.0.1